refact: removes splash screen rules and liability
