+++ src/pages/SplashPage/viewModel/interface.ts
@@ -1,0 +1,3 @@
+export interface ISplashViewModel {
+  handleCheckNextPage(): Promise<void>
+}
